Adds multi-architecture support
Enables building the Docker image for both linux/amd64 and linux/arm64 architectures.

This allows the image to be run on a wider range of devices and platforms.
It updates the Dockerfile to use build arguments for OS and architecture and adds platform specification to the build command in the Github workflow.